Begin reworking image and storage system
Right now, all of the image and storage login is in
`src/types/emote_image.rs`. This is a problem, since we might want to
create more robust image resizing later on (eg. different libraries for
different formats, exchanging libraries for different ones in the
future, etc.). We might also not want to use the local filesystem as a
storage medium, and maybe use something like S3 buckets.

As a result, I've created a new module `image` and `storage`. The
`image` module has structures and logic to create an abstracted API over
the resizer library, which can easily/trivially be called by
`emote_image.rs`. Everything is implemented except the actual library
calls from `libvips`, which should be easy to port from
`emote_image.rs`.

Furthermore, the `storage` module has a simple abstraction layer to
read/write bytes from a given UUID and extension name, along with
implementations for local and S3 storage.

Note that none of this has been tested yet, or integrated into the
current logic.